fix car racing pixel scale dtype (#1428)

+++ gym/envs/box2d/car_racing.py
@@ -370,8 +370,8 @@ class CarRacing(gym.Env, EzPickle):
             pixel_scale = 1
             if hasattr(win.context, '_nscontext'):
                 pixel_scale = win.context._nscontext.view().backingScaleFactor()  # pylint: disable=protected-access
-            VP_W = pixel_scale * WINDOW_W
-            VP_H = pixel_scale * WINDOW_H
+            VP_W = int(pixel_scale * WINDOW_W)
+            VP_H = int(pixel_scale * WINDOW_H)
 
         gl.glViewport(0, 0, VP_W, VP_H)
         t.enable()
